move UnitOfWork and PostRepository to the Repo folder
