--- CollabApp/CollabApp.mvc/Repo/IPostRepository.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using CollabApp.mvc.Models;
-
-namespace CollabApp.mvc.Repo
-{
-    public interface IPostRepository : IGenericRepository<Post>
-    {
-    }
-}
