+++ CollabApp/CollabApp.mvc/Repo/IUnitOfWork.cs
@@ -4,11 +4,12 @@ using System.Linq;
 using System.Threading.Tasks;
 using CollabApp.mvc.Repo;
 
-namespace CollabApp.mvc.Controllers
+namespace CollabApp.mvc.Repo
 {
     public interface IUnitOfWork
     {
         IPostRepository postRepository {get;}
+        IBoardRepository boardRepository {get;}
         Task CompleteAsync();
     }
 }
